Refactored components: SearchBar, CategoryGridItems
